refactor: move result custom dataclass methods to separate class
